yunwei/ops_server/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
idna==2.8
lxml==4.3.3
mysqlclient==1.4.2.post1
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.1
coreapi==2.3.3
daemonize==2.5.0
django-filter==2.2.0
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
jsonpath==0.82
pika==1.0.1
redis==3.2.1
setproctitle==1.1.10
uWSGI==2.0.18
ansible==2.7.8
ipip-ipdb==1.3.2
geoip2==3.0.0
rdbtools==0.1.14
python-lzf==0.2.4
kafka-python==2.0.1
